design finalized .Now adding css
